server/requirements.txt
Package           Version
----------------- -------
annotated-types   0.7.0
anyio             4.9.0
cffi              1.17.1
cryptography      45.0.5
ecdsa             0.19.1
fastapi           0.116.0
idna              3.10
passlib           1.7.4
pip               24.0
pyasn1            0.6.1
pycparser         2.22
pydantic          2.11.7
pydantic_core     2.33.2
python-jose       3.5.0
rsa               4.9.1
six               1.17.0
sniffio           1.3.1
starlette         0.46.2
typing_extensions 4.14.1
typing-inspection 0.4.1
